Add files via upload
An LLG solver is modelled on LTSpice and the netlist is synthesized which is then modified as desired for different input/test conditions. 
The spiceeditor class is used to modify the netlist during runtime; Simrunner class is used to run multiple instances of the same netlist.
The result is well analysed using various libraries and essential graphs are plotted.